# swagger.yaml
openapi: 3.0.0
info:
  title: FEMCS API
  description: Fire Extinguisher Management and Compliance System
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local API Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200': { description: List of customers }
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, phone]
              properties:
                fullName: { type: string }
                nationalId: { type: string }
                phone: { type: string }
                email: { type: string }
                address: { type: string }
                organizationName: { type: string }
      responses:
        '201': { description: Customer created }
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Customer data }
    put:
      tags: [Customers]
      summary: Update customer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Customer updated }
    delete:
      tags: [Customers]
      summary: Delete customer (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Customer deleted }
...